requirements.txt
# Binance Futures Testnet Trading Bot
python-dotenv>=1.0.0
httpx>=0.27.0
binance-futures-connector>=3.0.0
groq>=0.4.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
